Refactored methods that find available event times.
In an effort to make code cleaner and to avoid stack overflows, the main logic of EventServiceImpl.findAvailableEventTimes() was turned into a command.
