+++ src/main/java/com/knotslicer/server/adapters/rest/UserWithEventsResource.java
@@ -1,6 +1,6 @@
 package com.knotslicer.server.adapters.rest;
 
-import com.knotslicer.server.adapters.rest.linkgenerator.Invoker;
+import com.knotslicer.server.adapters.rest.linkgenerator.LinkInvoker;
 import com.knotslicer.server.adapters.rest.linkgenerator.LinkReceiver;
 import com.knotslicer.server.ports.interactor.WithChildren;
 import com.knotslicer.server.adapters.rest.linkgenerator.linkcommands.LinkCommand;
@@ -47,9 +47,9 @@ public class UserWithEventsResource implements UserWithChildrenResource {
                 .build();
     }
     private URI addLinks(LinkCommand<UserLightDto> linkCommand) {
-        Invoker invoker =
-                userWithEventsLinkCreator.createInvoker(linkCommand);
-        return invoker.executeCommand();
+        LinkInvoker linkInvoker =
+                userWithEventsLinkCreator.createLinkInvoker(linkCommand);
+        return linkInvoker.executeCommand();
     }
     @Inject
     public UserWithEventsResource(@ProcessAs(ProcessType.USER) @WithChildren(ProcessType.EVENT)
