+++ src/main/java/com/knotslicer/server/adapters/rest/linkgenerator/LinkInvoker.java
@@ -2,6 +2,6 @@ package com.knotslicer.server.adapters.rest.linkgenerator;
 
 import java.net.URI;
 
-public interface Invoker {
+public interface LinkInvoker {
     URI executeCommand();
 }
