+++ src/main/java/com/knotslicer/server/ports/interactor/InteractorCommand.java
@@ -1,0 +1,5 @@
+package com.knotslicer.server.ports.interactor;
+
+public interface InteractorCommand {
+    Boolean execute();
+}
