+++ src/main/java/com/knotslicer/server/ports/interactor/InteractorCommandInvoker.java
@@ -1,0 +1,5 @@
+package com.knotslicer.server.ports.interactor;
+
+public interface InteractorCommandInvoker {
+    Boolean executeCommand();
+}
