+++ src/main/java/com/knotslicer/server/ports/interactor/InteractorCommandInvokerImpl.java
@@ -1,0 +1,12 @@
+package com.knotslicer.server.ports.interactor;
+
+public class InteractorCommandInvokerImpl implements InteractorCommandInvoker {
+    private InteractorCommand interactorCommand;
+    public InteractorCommandInvokerImpl(InteractorCommand interactorCommand) {
+        this.interactorCommand = interactorCommand;
+    }
+    @Override
+    public Boolean executeCommand() {
+        return interactorCommand.execute();
+    }
+}
